fix(codegen): keep the operand stack balanced in tag-based virtual dispatch
Two defects in emitVirtualMethodDispatchByTag, both of which made lit's
implementation module fail WebAssembly.compile. An invalid implementation
costs far more than the one package: the upstream-suite runner halves and
retries any unit that fails validation (depth <= 6, so up to 64 full-bundle
recompiles for a single test file), which is a large part of lit's
~28-minute measurement row.

1. PARTIAL EMISSION ON BAIL-OUT. The function compiled the receiver into
   fctx.body and only then checked it was ref-typed. Returning undefined
   means "I emitted nothing, use the static path", so for an EXTERNREF
   receiver — what `class extends HTMLElement` produces — the receiver push
   stayed on the stack, the static path pushed it again, and one externref
   outlived the call. Inside `t && this.k(1)` that surviving value became
   the value of an arm typed i32: "type error in fallthru[0] (expected i32,
   got externref)". Both post-emission bail-outs now roll back through the
   existing speculative snapshot.

2. TAG LOAD THROUGH THE WRONG STRUCT. The cascade read `struct.get $base`
   from a local typed with the receiver's own struct. These class structs
   are not declared in a Wasm subtype relation (`$__anonClass_0` is emitted
   as a bare `(sub (struct …))`), so that is a validation error, not a
   widening. It fires when the method is compiled as a synthesised
   per-subclass copy, where `this` is the copy's struct while baseClassName
   still names the original — lit's y_get_updateComplete failed with
   "struct.get[0] expected type (ref null 54), found local.tee of type
   (ref null 44)". Field 0 is __tag in every class struct on this path, so
   the read now goes through the receiver's own type, and a receiver that
   is not a known class struct bails out (safely, given fix 1).

The emitter moves to src/codegen/expressions/virtual-dispatch.ts. calls.ts
is a god file under the #3102 LOC budget with 3 lines of headroom, and the
gate's own advice is to add code to the subsystem module rather than the
driver; calls.ts re-exports the entry point so no call site changes. The
move is verbatim apart from the two fixes.

Measured: lit-html, @lit/reactive-element and lit-element each compiled
INVALID before and each VALIDATES now. The three-package bundle still
fails on a third, unrelated defect (__anon_42_set: i32.eq on an externref
struct.get) that only appears once lit/decorators.js is included — not
addressed here.

Regression tests fail 4/5 without this change. The fifth runs the compiled
module and asserts the cascade still selects the overriding implementation,
so a stack fix that broke dispatch would not pass.

No regressions: required guard suite (225 passed), the 9 class/inheritance
suites (63), the 19 class-related equivalence files (153), typecheck, and
the LOC/function budget gates.
tests/equivalence/optional-direct-closure-call.test.ts fails with AND
without this change — pre-existing on main, reported separately.

Pre-commit checklist completed ✓
